% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{crp_list_segments}
\alias{crp_list_segments}
\title{List segments available in a basin}
\usage{
crp_list_segments(basin)
}
\arguments{
\item{basin}{Basin number (1-25). See \code{\link[=crp_basins]{crp_basins()}}.}
}
\value{
A tibble with \code{row}, \code{segment_id}, \code{segment_desc}.
}
\description{
List segments available in a basin
}
