% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{crp_segments}
\alias{crp_segments}
\title{Extract the segment table for the selected basin}
\usage{
crp_segments(form)
}
\arguments{
\item{form}{A \code{crp_form} from \code{\link[=crp_select_basin]{crp_select_basin()}}.}
}
\value{
A tibble with columns \code{row}, \code{segment_id}, \code{segment_desc}.
}
\description{
Extract the segment table for the selected basin
}
